Extract the testing configuration from group testing results

Extract the testing configuration from objects of class "OTC" returned by OTC1 (OTC1) or OTC2 (OTC2).

Usage

# S3 method for class 'OTC'
Config(object, n = 5, top.overall = FALSE, ...)

Arguments

object

An object of class "OTC", from which the testing configuration is to be extracted.

n

Number of testing configurations.

top.overall

logical; if TRUE, best overall testing configurations; if FALSE, best testing configurations by initial group size

...

currently not used.

Value

A data frame providing the best testing configurations.

Author

Christopher R. Bilder

Examples

res1 <- OTC1(algorithm = "D3", p = 0.05, Se = 0.99, Sp = 0.99,
             group.sz = 3:15, obj.fn = "ET")
#> Initial Group Size = 3
#> Initial Group Size = 4
#> Initial Group Size = 5
#> Initial Group Size = 6
#> Initial Group Size = 7
#> Initial Group Size = 8
#> Initial Group Size = 9
#> Initial Group Size = 10
#> Initial Group Size = 11
#> Initial Group Size = 12
#> Initial Group Size = 13
#> Initial Group Size = 14
#> Initial Group Size = 15
#> 
#>  Number of minutes running:  0 
#>  
Config(res1)
#>    I  config     ET  value    PSe    PSp   PPPV   PNPV
#> 1  9   3,3,3 3.3960 0.3773 0.9703 0.9990 0.9812 0.9984
#> 2 12 3,3,3,3 4.5575 0.3798 0.9703 0.9990 0.9810 0.9984
#> 3 10   4,3,3 3.8001 0.3800 0.9703 0.9988 0.9778 0.9984
#> 4 11   4,4,3 4.2002 0.3818 0.9703 0.9987 0.9751 0.9984
#> 5 13 4,3,3,3 4.9762 0.3828 0.9703 0.9989 0.9784 0.9984