Extract the expected number of tests from testing configuration results

Extract the expected number of tests from objects of class "Sterrett" returned by Sterrett (Sterrett).

# S3 method for Sterrett
ExpTests(object, ...)

Arguments

object

An object of class "Sterrett", from which the expected number of tests is to be extracted.

...

Additional arguments to be passed to ExpTests (e.g., digits to be passed to round for appropriate rounding).

Value

A data frame containing the columns:

ExpTests

the expected number of tests required to decode all individuals in the algorithm.

ExpTestsPerIndividual

the expected number of tests per individual.

PercentReductionTests

The percent reduction in the number of tests; 100 * (1 - ExpTestsPerIndividual).

PercentIncreaseTestCap

The percent increase in testing capacity when the algorithm is applied to a continuous stream of specimens; 100 * (1/ExpTestsPerIndividual - 1).

References

Bilder, C., Iwen, P., Abdalhamid, B., Tebbs, J., McMahan, C. (2020). “Tests in short supply? Try group testing.” Significance, 17, 15.

Author

Christopher R. Bilder

Examples

set.seed(1231)
p.vec1 <- rbeta(n = 8, shape1 = 1, shape2 = 10)
save.it1 <- Sterrett(p = p.vec1, Sp = 0.90, Se = 0.95)
ExpTests(save.it1)
#>    Method ExpTests ExpTestsPerIndividual PercentReductionTests
#> 1    1SIS   3.9808                0.4976                 50.24
#> 2    2SIS   3.6693                0.4587                 54.13
#> 3     FIS   3.6123                0.4515                 54.85
#> 4 Dorfman   5.0932                0.6366                 36.34
#>   PercentIncreaseTestCap
#> 1                 100.96
#> 2                 118.01
#> 3                 121.48
#> 4                  57.08