Extract the expected number of tests from testing configuration results

Extract the expected number of tests from objects of class "TOD" returned by TOD (TOD).

Usage

# S3 method for class 'TOD'
ExpTests(object, ...)

Arguments

object

An object of class "TOD", from which the expected number of tests is to be extracted.

...

Additional arguments to be passed to ExpTests (e.g., digits to be passed to round for appropriate rounding).

Value

A data frame containing the columns:

ExpTests

the expected number of tests required to decode all individuals in the algorithm.

ExpTestsPerIndividual

the expected number of tests per individual.

PercentReductionTests

The percent reduction in the number of tests; 100 * (1 - ExpTestsPerIndividual).

PercentIncreaseTestCap

The percent increase in testing capacity when the algorithm is applied to a continuous stream of specimens; 100 * (1/ExpTestsPerIndividual - 1).

References

Bilder, C., Iwen, P., Abdalhamid, B., Tebbs, J., McMahan, C. (2020). “Tests in short supply? Try group testing.” Significance, 17, 15.

Author

Christopher R. Bilder

Examples

set.seed(1002)
p.vec <- expectOrderBeta(p = 0.01, alpha = 2, size = 20)
save.it1 <- TOD(p = p.vec, Se = 0.95, Sp = 0.95, max = 5, threshold = 0.015)
ExpTests(save.it1)
#>   ExpTests ExpTestsPerIndividual PercentReductionTests PercentIncreaseTestCap
#> 1   9.2189                0.4609                 53.91                 116.97