Extract probability mass function (PMF) from group testing results

Extract the probability mass function from group testing results for the Sterrett algorithm (objects of class "Sterrett" returned by Sterrett).

Usage

# S3 method for class 'Sterrett'
pmf(object, ...)

Arguments

object

An object of class "Sterrett", created by Sterrett, from which the PMF is to be extracted.

...

currently not used.

Value

Data frame containing the probability mass function extracted from the object object.

Author

Brianna D. Hitt

Examples

set.seed(1231)
p.vec <- rbeta(n = 8, shape1 = 1, shape2 = 10)
res <- Sterrett(p = p.vec, Sp = 0.90, Se = 0.95)
pmf(res)
#>    tests one.stage.IS.PMF two.stage.IS.PMF  full.IS.PMF Dorfman.PMF
#> 1      1       0.48835471       0.48835471 4.883547e-01   0.4883547
#> 2      3       0.09599437       0.09599437 9.599437e-02          NA
#> 3      4       0.07394624       0.07394624 7.394624e-02          NA
#> 4      5       0.04488706       0.06729731 6.729731e-02          NA
#> 5      6       0.03972361       0.06516755 6.516755e-02          NA
#> 6      7       0.02005756       0.05073281 5.444685e-02          NA
#> 7      8       0.01199658       0.03078615 3.887818e-02          NA
#> 8      9       0.04084418       0.05308898 6.080128e-02   0.5116453
#> 9     10       0.18419568       0.03045444 3.711694e-02          NA
#> 10    11               NA       0.04417744 1.426534e-02          NA
#> 11    12               NA               NA 3.262503e-03          NA
#> 12    13               NA               NA 4.362291e-04          NA
#> 13    14               NA               NA 3.151575e-05          NA
#> 14    15               NA               NA 9.742431e-07          NA