Predict method for gtReg

Obtains predictions for individual observations and optionally computes the standard errors of those predictions from objects of class "gtReg" returned by gtReg.

# S3 method for gtReg
predict(
  object,
  newdata,
  type = c("link", "response"),
  se.fit = FALSE,
  conf.level = NULL,
  na.action = na.pass,
  ...
)

Arguments

object

a fitted object of class "gtReg".

newdata

an optional data frame in which to look for variables with which to predict. If omitted, the fitted linear predictors are used.

type

the type of prediction required. The "link" option is on the scale of the linear predictors. The "response" option is on the scale of the response variable. Thus, for the logit model, the "link" predictions are of log-odds (probabilities on the logit scale) and type = "response" gives the predicted probabilities.

se.fit

a logical value indicating whether standard errors are required.

conf.level

the confidence level of the interval for the predicted values.

na.action

a function determining what should be done with missing values in newdata. The default is to predict NA.

...

currently not used.

Value

If se = FALSE, a vector or matrix of predictions. If se = TRUE, a list containing:

fit

predictions.

se.fit

estimated standard errors.

lower

the lower bound of the confidence interval, if calculated (i.e., conf.level is specified).

upper

the upper bound of the confidence interval, if calculated (i.e., conf.level is specified).

Details

If newdata is omitted, the predictions are based on the data used for the fit. When newdata is present and contains missing values, how the missing values will be dealt with is determined by the na.action argument. In this case, if na.action = na.omit, omitted cases will not appear, whereas if na.action = na.exclude, omitted cases will appear (in predictions and standard errors) with value NA.

Author

Boan Zhang

Examples

data(hivsurv)
fit1 <- gtReg(formula = groupres ~ AGE + EDUC.,
              data = hivsurv, groupn = gnum,
              sens = 0.9, spec = 0.9,
              linkf = "logit", method = "V")
pred.data <- data.frame(AGE = c(15, 25, 30),
                        EDUC. = c(1, 3, 2))
predict(object = fit1, newdata = pred.data,
        type = "link", se.fit = TRUE)
#>      fit se.fit
#> 1 -3.145 0.8906
#> 2 -2.058 0.3168
#> 3 -3.171 0.7505
predict(object = fit1, newdata = pred.data,
        type = "response", se.fit = TRUE,
        conf.level = 0.9)
#>       fit  se.fit    lower  upper
#> 1 0.04130 0.03527 0.009858 0.1571
#> 2 0.11320 0.03181 0.070500 0.1770
#> 3 0.04027 0.02901 0.012060 0.1260
predict(object = fit1, type = "response",
        se.fit = TRUE, conf.level = 0.9)
#>          fit  se.fit     lower  upper
#> 1   0.268600 0.15980 0.0878300 0.5834
#> 2   0.085220 0.04699 0.0334100 0.2007
#> 3   0.037020 0.03001 0.0095340 0.1331
#> 4   0.080890 0.04066 0.0345600 0.1779
#> 5   0.035040 0.02798 0.0092260 0.1241
#> 6   0.069050 0.02718 0.0356800 0.1295
#> 7   0.050060 0.02446 0.0221200 0.1094
#> 8   0.030590 0.03260 0.0051470 0.1614
#> 9   0.080890 0.04066 0.0345600 0.1779
#> 10  0.076750 0.03521 0.0354100 0.1584
#> 11  0.125200 0.03527 0.0777200 0.1955
#> 12  0.113200 0.03181 0.0705000 0.1770
#> 13  0.113200 0.03181 0.0705000 0.1770
#> 14  0.052840 0.02348 0.0251400 0.1077
#> 15  0.107600 0.03283 0.0643400 0.1747
#> 16  0.097180 0.03783 0.0503000 0.1795
#> 17  0.092300 0.04092 0.0435500 0.1851
#> 18  0.013600 0.01954 0.0012530 0.1315
#> 19  0.030590 0.03260 0.0051470 0.1614
#> 20  0.012150 0.01906 0.0009019 0.1436
#> 21  0.060560 0.05953 0.0114000 0.2650
#> 22  0.031390 0.02493 0.0083390 0.1110
#> 23  0.145200 0.05540 0.0753300 0.2614
#> 24  0.131600 0.04015 0.0783400 0.2126
#> 25  0.131600 0.04015 0.0783400 0.2126
#> 26  0.131600 0.04015 0.0783400 0.2126
#> 27  0.119100 0.03250 0.0751200 0.1837
#> 28  0.113200 0.03181 0.0705000 0.1770
#> 29  0.052840 0.02348 0.0251400 0.1077
#> 30  0.268600 0.15980 0.0878300 0.5834
#> 31  0.131600 0.04015 0.0783400 0.2126
#> 32  0.125200 0.03527 0.0777200 0.1955
#> 33  0.050060 0.02446 0.0221200 0.1094
#> 34  0.085220 0.04699 0.0334100 0.2007
#> 35  0.085220 0.04699 0.0334100 0.2007
#> 36  0.076750 0.03521 0.0354100 0.1584
#> 37  0.035040 0.02798 0.0092260 0.1241
#> 38  0.062080 0.02322 0.0332100 0.1131
#> 39  0.023790 0.02146 0.0053020 0.1003
#> 40  0.102300 0.03501 0.0573800 0.1758
#> 41  0.092300 0.04092 0.0435500 0.1851
#> 42  0.042530 0.02793 0.0141700 0.1207
#> 43  0.019610 0.03257 0.0012320 0.2450
#> 44  0.041300 0.03527 0.0098580 0.1571
#> 45  0.175800 0.10200 0.0627600 0.4045
#> 46  0.080890 0.04066 0.0345600 0.1779
#> 47  0.035040 0.02798 0.0092260 0.1241
#> 48  0.031390 0.02493 0.0083390 0.1110
#> 49  0.145200 0.05540 0.0753300 0.2614
#> 50  0.065480 0.02470 0.0348200 0.1198
#> 51  0.062080 0.02322 0.0332100 0.1131
#> 52  0.125200 0.03527 0.0777200 0.1955
#> 53  0.055770 0.02281 0.0281500 0.1075
#> 54  0.023790 0.02146 0.0053020 0.1003
#> 55  0.080890 0.04066 0.0345600 0.1779
#> 56  0.159900 0.07639 0.0694800 0.3266
#> 57  0.031390 0.02493 0.0083390 0.1110
#> 58  0.031390 0.02493 0.0083390 0.1110
#> 59  0.062080 0.02322 0.0332100 0.1131
#> 60  0.062080 0.02322 0.0332100 0.1131
#> 61  0.018000 0.02044 0.0027280 0.1094
#> 62  0.016090 0.02012 0.0020180 0.1169
#> 63  0.076750 0.03521 0.0354100 0.1584
#> 64  0.031390 0.02493 0.0083390 0.1110
#> 65  0.069050 0.02718 0.0356800 0.1295
#> 66  0.058840 0.02265 0.0309200 0.1091
#> 67  0.023790 0.02146 0.0053020 0.1003
#> 68  0.018000 0.02044 0.0027280 0.1094
#> 69  0.159900 0.07639 0.0694800 0.3266
#> 70  0.065480 0.02470 0.0348200 0.1198
#> 71  0.065480 0.02470 0.0348200 0.1198
#> 72  0.119100 0.03250 0.0751200 0.1837
#> 73  0.023790 0.02146 0.0053020 0.1003
#> 74  0.138200 0.04695 0.0773800 0.2347
#> 75  0.131600 0.04015 0.0783400 0.2126
#> 76  0.058840 0.02265 0.0309200 0.1091
#> 77  0.113200 0.03181 0.0705000 0.1770
#> 78  0.107600 0.03283 0.0643400 0.1747
#> 79  0.047420 0.02560 0.0192200 0.1122
#> 80  0.092300 0.04092 0.0435500 0.1851
#> 81  0.060560 0.05953 0.0114000 0.2650
#> 82  0.054390 0.06151 0.0079810 0.2914
#> 83  0.072810 0.03070 0.0358300 0.1423
#> 84  0.125200 0.03527 0.0777200 0.1955
#> 85  0.119100 0.03250 0.0751200 0.1837
#> 86  0.047420 0.02560 0.0192200 0.1122
#> 87  0.080890 0.04066 0.0345600 0.1779
#> 88  0.078940 0.04980 0.0270300 0.2091
#> 89  0.039110 0.03243 0.0097440 0.1441
#> 90  0.080890 0.04066 0.0345600 0.1779
#> 91  0.076750 0.03521 0.0354100 0.1584
#> 92  0.072810 0.03070 0.0358300 0.1423
#> 93  0.069050 0.02718 0.0356800 0.1295
#> 94  0.058840 0.02265 0.0309200 0.1091
#> 95  0.058840 0.02265 0.0309200 0.1091
#> 96  0.055770 0.02281 0.0281500 0.1075
#> 97  0.107600 0.03283 0.0643400 0.1747
#> 98  0.102300 0.03501 0.0573800 0.1758
#> 99  0.044910 0.02678 0.0165600 0.1161
#> 100 0.044910 0.02678 0.0165600 0.1161
#> 101 0.087630 0.04404 0.0373700 0.1920
#> 102 0.063880 0.05816 0.0135900 0.2525
#> 103 0.037020 0.03001 0.0095340 0.1331
#> 104 0.069050 0.02718 0.0356800 0.1295
#> 105 0.131600 0.04015 0.0783400 0.2126
#> 106 0.023790 0.02146 0.0053020 0.1003
#> 107 0.113200 0.03181 0.0705000 0.1770
#> 108 0.050060 0.02446 0.0221200 0.1094
#> 109 0.097180 0.03783 0.0503000 0.1795
#> 110 0.040270 0.02901 0.0120600 0.1260
#> 111 0.013600 0.01954 0.0012530 0.1315
#> 112 0.010260 0.01820 0.0005438 0.1650
#> 113 0.069050 0.02718 0.0356800 0.1295
#> 114 0.268600 0.15980 0.0878300 0.5834
#> 115 0.052840 0.02348 0.0251400 0.1077
#> 116 0.102300 0.03501 0.0573800 0.1758
#> 117 0.097180 0.03783 0.0503000 0.1795
#> 118 0.089770 0.05418 0.0320700 0.2269
#> 119 0.119100 0.03250 0.0751200 0.1837
#> 120 0.052840 0.02348 0.0251400 0.1077
#> 121 0.016090 0.02012 0.0020180 0.1169
#> 122 0.167700 0.08868 0.0661500 0.3642
#> 123 0.033170 0.02630 0.0088240 0.1168
#> 124 0.152400 0.06527 0.0726000 0.2922
#> 125 0.069050 0.02718 0.0356800 0.1295
#> 126 0.055770 0.02281 0.0281500 0.1075
#> 127 0.236300 0.13000 0.0864800 0.5029
#> 128 0.107600 0.03283 0.0643400 0.1747
#> 129 0.107600 0.03283 0.0643400 0.1747
#> 130 0.050060 0.02446 0.0221200 0.1094
#> 131 0.087630 0.04404 0.0373700 0.1920
#> 132 0.089770 0.05418 0.0320700 0.2269
#> 133 0.085220 0.04699 0.0334100 0.2007
#> 134 0.076750 0.03521 0.0354100 0.1584
#> 135 0.069050 0.02718 0.0356800 0.1295
#> 136 0.055770 0.02281 0.0281500 0.1075
#> 137 0.113200 0.03181 0.0705000 0.1770
#> 138 0.022500 0.02117 0.0047040 0.1008
#> 139 0.047420 0.02560 0.0192200 0.1122
#> 140 0.018000 0.02044 0.0027280 0.1094
#> 141 0.010860 0.01850 0.0006446 0.1574
#> 142 0.145200 0.05540 0.0753300 0.2614
#> 143 0.019030 0.02060 0.0031520 0.1064
#> 144 0.018000 0.02044 0.0027280 0.1094
#> 145 0.069050 0.02718 0.0356800 0.1295
#> 146 0.145200 0.05540 0.0753300 0.2614
#> 147 0.125200 0.03527 0.0777200 0.1955
#> 148 0.058840 0.02265 0.0309200 0.1091
#> 149 0.246800 0.13840 0.0878200 0.5272
#> 150 0.119100 0.03250 0.0751200 0.1837
#> 151 0.216400 0.11790 0.0808800 0.4643
#> 152 0.230900 0.20270 0.0438800 0.6625
#> 153 0.080890 0.04066 0.0345600 0.1779
#> 154 0.031390 0.02493 0.0083390 0.1110
#> 155 0.062080 0.02322 0.0332100 0.1131
#> 156 0.058840 0.02265 0.0309200 0.1091
#> 157 0.119100 0.03250 0.0751200 0.1837
#> 158 0.113200 0.03181 0.0705000 0.1770
#> 159 0.113200 0.03181 0.0705000 0.1770
#> 160 0.092300 0.04092 0.0435500 0.1851
#> 161 0.092300 0.04092 0.0435500 0.1851
#> 162 0.083180 0.04703 0.0318600 0.2001
#> 163 0.016090 0.02012 0.0020180 0.1169
#> 164 0.074900 0.05232 0.0228500 0.2190
#> 165 0.074900 0.05232 0.0228500 0.2190
#> 166 0.067370 0.05650 0.0161900 0.2407
#> 167 0.039110 0.03243 0.0097440 0.1441
#> 168 0.076750 0.03521 0.0354100 0.1584
#> 169 0.031390 0.02493 0.0083390 0.1110
#> 170 0.062080 0.02322 0.0332100 0.1131
#> 171 0.107600 0.03283 0.0643400 0.1747
#> 172 0.044910 0.02678 0.0165600 0.1161
#> 173 0.020130 0.02076 0.0036240 0.1040
#> 174 0.020130 0.02076 0.0036240 0.1040
#> 175 0.092300 0.04092 0.0435500 0.1851
#> 176 0.172000 0.11090 0.0545400 0.4279
#> 177 0.172000 0.11090 0.0545400 0.4279
#> 178 0.085220 0.04699 0.0334100 0.2007
#> 179 0.257500 0.14840 0.0882300 0.5542
#> 180 0.119100 0.03250 0.0751200 0.1837
#> 181 0.107600 0.03283 0.0643400 0.1747
#> 182 0.047420 0.02560 0.0192200 0.1122
#> 183 0.206900 0.11410 0.0767100 0.4503
#> 184 0.047420 0.02560 0.0192200 0.1122
#> 185 0.197700 0.11180 0.0717800 0.4398
#> 186 0.013600 0.01954 0.0012530 0.1315
#> 187 0.039110 0.03243 0.0097440 0.1441
#> 188 0.028100 0.02300 0.0071860 0.1035
#> 189 0.125200 0.03527 0.0777200 0.1955
#> 190 0.020130 0.02076 0.0036240 0.1040
#> 191 0.076750 0.03521 0.0354100 0.1584
#> 192 0.031390 0.02493 0.0083390 0.1110
#> 193 0.069050 0.02718 0.0356800 0.1295
#> 194 0.138200 0.04695 0.0773800 0.2347
#> 195 0.138200 0.04695 0.0773800 0.2347
#> 196 0.062080 0.02322 0.0332100 0.1131
#> 197 0.055770 0.02281 0.0281500 0.1075
#> 198 0.050060 0.02446 0.0221200 0.1094
#> 199 0.047420 0.02560 0.0192200 0.1122
#> 200 0.021280 0.02095 0.0041420 0.1021
#> 201 0.102300 0.03501 0.0573800 0.1758
#> 202 0.097180 0.03783 0.0503000 0.1795
#> 203 0.040270 0.02901 0.0120600 0.1260
#> 204 0.038120 0.02998 0.0102300 0.1320
#> 205 0.078940 0.04980 0.0270300 0.2091
#> 206 0.025910 0.03330 0.0030270 0.1890
#> 207 0.072810 0.03070 0.0358300 0.1423
#> 208 0.138200 0.04695 0.0773800 0.2347
#> 209 0.026590 0.02234 0.0065590 0.1015
#> 210 0.052840 0.02348 0.0251400 0.1077
#> 211 0.023790 0.02146 0.0053020 0.1003
#> 212 0.044910 0.02678 0.0165600 0.1161
#> 213 0.087630 0.04404 0.0373700 0.1920
#> 214 0.035040 0.02798 0.0092260 0.1241
#> 215 0.076750 0.03521 0.0354100 0.1584
#> 216 0.031390 0.02493 0.0083390 0.1110
#> 217 0.018000 0.02044 0.0027280 0.1094
#> 218 0.069050 0.02718 0.0356800 0.1295
#> 219 0.138200 0.04695 0.0773800 0.2347
#> 220 0.044910 0.02678 0.0165600 0.1161
#> 221 0.152400 0.06527 0.0726000 0.2922
#> 222 0.050060 0.02446 0.0221200 0.1094
#> 223 0.072810 0.03070 0.0358300 0.1423
#> 224 0.069050 0.02718 0.0356800 0.1295
#> 225 0.069050 0.02718 0.0356800 0.1295
#> 226 0.031390 0.02493 0.0083390 0.1110
#> 227 0.065480 0.02470 0.0348200 0.1198
#> 228 0.257500 0.14840 0.0882300 0.5542
#> 229 0.018000 0.02044 0.0027280 0.1094
#> 230 0.040270 0.02901 0.0120600 0.1260
#> 231 0.028950 0.03295 0.0043170 0.1701
#> 232 0.028950 0.03295 0.0043170 0.1701
#> 233 0.011490 0.01879 0.0007630 0.1503
#> 234 0.085220 0.04699 0.0334100 0.2007
#> 235 0.085220 0.04699 0.0334100 0.2007
#> 236 0.028100 0.02300 0.0071860 0.1035
#> 237 0.216400 0.11790 0.0808800 0.4643
#> 238 0.206900 0.11410 0.0767100 0.4503
#> 239 0.016090 0.02012 0.0020180 0.1169
#> 240 0.078940 0.04980 0.0270300 0.2091
#> 241 0.074900 0.05232 0.0228500 0.2190
#> 242 0.080890 0.04066 0.0345600 0.1779
#> 243 0.080890 0.04066 0.0345600 0.1779
#> 244 0.131600 0.04015 0.0783400 0.2126
#> 245 0.062080 0.02322 0.0332100 0.1131
#> 246 0.083180 0.04703 0.0318600 0.2001
#> 247 0.055770 0.02281 0.0281500 0.1075
#> 248 0.036090 0.03082 0.0086410 0.1385
#> 249 0.072810 0.03070 0.0358300 0.1423
#> 250 0.069050 0.02718 0.0356800 0.1295
#> 251 0.138200 0.04695 0.0773800 0.2347
#> 252 0.065480 0.02470 0.0348200 0.1198
#> 253 0.246800 0.13840 0.0878200 0.5272
#> 254 0.055770 0.02281 0.0281500 0.1075
#> 255 0.074900 0.05232 0.0228500 0.2190
#> 256 0.279900 0.17270 0.0867200 0.6141
#> 257 0.145200 0.05540 0.0753300 0.2614
#> 258 0.065480 0.02470 0.0348200 0.1198
#> 259 0.028100 0.02300 0.0071860 0.1035
#> 260 0.062080 0.02322 0.0332100 0.1131
#> 261 0.131600 0.04015 0.0783400 0.2126
#> 262 0.236300 0.13000 0.0864800 0.5029
#> 263 0.020130 0.02076 0.0036240 0.1040
#> 264 0.019030 0.02060 0.0031520 0.1064
#> 265 0.018000 0.02044 0.0027280 0.1094
#> 266 0.076750 0.03521 0.0354100 0.1584
#> 267 0.031390 0.02493 0.0083390 0.1110
#> 268 0.119100 0.03250 0.0751200 0.1837
#> 269 0.052840 0.02348 0.0251400 0.1077
#> 270 0.062080 0.02322 0.0332100 0.1131
#> 271 0.050060 0.02446 0.0221200 0.1094
#> 272 0.036090 0.03082 0.0086410 0.1385
#> 273 0.072810 0.03070 0.0358300 0.1423
#> 274 0.119100 0.03250 0.0751200 0.1837
#> 275 0.078940 0.04980 0.0270300 0.2091
#> 276 0.145200 0.05540 0.0753300 0.2614
#> 277 0.031390 0.02493 0.0083390 0.1110
#> 278 0.028100 0.02300 0.0071860 0.1035
#> 279 0.055770 0.02281 0.0281500 0.1075
#> 280 0.015210 0.01995 0.0017260 0.1213
#> 281 0.012850 0.01931 0.0010640 0.1373
#> 282 0.008664 0.01719 0.0003245 0.1905
#> 283 0.069050 0.02718 0.0356800 0.1295
#> 284 0.014380 0.01976 0.0014730 0.1262
#> 285 0.138200 0.04695 0.0773800 0.2347
#> 286 0.131600 0.04015 0.0783400 0.2126
#> 287 0.125200 0.03527 0.0777200 0.1955
#> 288 0.021280 0.02095 0.0041420 0.1021
#> 289 0.067370 0.05650 0.0161900 0.2407
#> 290 0.023790 0.02146 0.0053020 0.1003
#> 291 0.020130 0.02076 0.0036240 0.1040
#> 292 0.069050 0.02718 0.0356800 0.1295
#> 293 0.062080 0.02322 0.0332100 0.1131
#> 294 0.131600 0.04015 0.0783400 0.2126
#> 295 0.047420 0.02560 0.0192200 0.1122
#> 296 0.102300 0.03501 0.0573800 0.1758
#> 297 0.092300 0.04092 0.0435500 0.1851
#> 298 0.083180 0.04703 0.0318600 0.2001
#> 299 0.033170 0.02630 0.0088240 0.1168
#> 300 0.055770 0.02281 0.0281500 0.1075
#> 301 0.050060 0.02446 0.0221200 0.1094
#> 302 0.036090 0.03082 0.0086410 0.1385
#> 303 0.039110 0.03243 0.0097440 0.1441
#> 304 0.159900 0.07639 0.0694800 0.3266
#> 305 0.033170 0.02630 0.0088240 0.1168
#> 306 0.062080 0.02322 0.0332100 0.1131
#> 307 0.102300 0.03501 0.0573800 0.1758
#> 308 0.044910 0.02678 0.0165600 0.1161
#> 309 0.080890 0.04066 0.0345600 0.1779
#> 310 0.145200 0.05540 0.0753300 0.2614
#> 311 0.029700 0.02385 0.0077860 0.1067
#> 312 0.065480 0.02470 0.0348200 0.1198
#> 313 0.028100 0.02300 0.0071860 0.1035
#> 314 0.026590 0.02234 0.0065590 0.1015
#> 315 0.058840 0.02265 0.0309200 0.1091
#> 316 0.050060 0.02446 0.0221200 0.1094
#> 317 0.197700 0.11180 0.0717800 0.4398
#> 318 0.015210 0.01995 0.0017260 0.1213
#> 319 0.141900 0.11700 0.0329400 0.4453
#> 320 0.069050 0.02718 0.0356800 0.1295
#> 321 0.131600 0.04015 0.0783400 0.2126
#> 322 0.062080 0.02322 0.0332100 0.1131
#> 323 0.055770 0.02281 0.0281500 0.1075
#> 324 0.197700 0.11180 0.0717800 0.4398
#> 325 0.037020 0.03001 0.0095340 0.1331
#> 326 0.159900 0.07639 0.0694800 0.3266
#> 327 0.033170 0.02630 0.0088240 0.1168
#> 328 0.062080 0.02322 0.0332100 0.1131
#> 329 0.062080 0.02322 0.0332100 0.1131
#> 330 0.025150 0.02184 0.0059250 0.1005
#> 331 0.119100 0.03250 0.0751200 0.1837
#> 332 0.022500 0.02117 0.0047040 0.1008
#> 333 0.040270 0.02901 0.0120600 0.1260
#> 334 0.013600 0.01954 0.0012530 0.1315
#> 335 0.037020 0.03001 0.0095340 0.1331
#> 336 0.076750 0.03521 0.0354100 0.1584
#> 337 0.069050 0.02718 0.0356800 0.1295
#> 338 0.131600 0.04015 0.0783400 0.2126
#> 339 0.026590 0.02234 0.0065590 0.1015
#> 340 0.125200 0.03527 0.0777200 0.1955
#> 341 0.246800 0.13840 0.0878200 0.5272
#> 342 0.102300 0.03501 0.0573800 0.1758
#> 343 0.007312 0.01609 0.0001921 0.2202
#> 344 0.085220 0.04699 0.0334100 0.2007
#> 345 0.175800 0.10200 0.0627600 0.4045
#> 346 0.080890 0.04066 0.0345600 0.1779
#> 347 0.035040 0.02798 0.0092260 0.1241
#> 348 0.035040 0.02798 0.0092260 0.1241
#> 349 0.076750 0.03521 0.0354100 0.1584
#> 350 0.031390 0.02493 0.0083390 0.1110
#> 351 0.069050 0.02718 0.0356800 0.1295
#> 352 0.025150 0.02184 0.0059250 0.1005
#> 353 0.025150 0.02184 0.0059250 0.1005
#> 354 0.023790 0.02146 0.0053020 0.1003
#> 355 0.107600 0.03283 0.0643400 0.1747
#> 356 0.020130 0.02076 0.0036240 0.1040
#> 357 0.011490 0.01879 0.0007630 0.1503
#> 358 0.080890 0.04066 0.0345600 0.1779
#> 359 0.076750 0.03521 0.0354100 0.1584
#> 360 0.076750 0.03521 0.0354100 0.1584
#> 361 0.055770 0.02281 0.0281500 0.1075
#> 362 0.159900 0.07639 0.0694800 0.3266
#> 363 0.072810 0.03070 0.0358300 0.1423
#> 364 0.069050 0.02718 0.0356800 0.1295
#> 365 0.065480 0.02470 0.0348200 0.1198
#> 366 0.131600 0.04015 0.0783400 0.2126
#> 367 0.125200 0.03527 0.0777200 0.1955
#> 368 0.058840 0.02265 0.0309200 0.1091
#> 369 0.246800 0.13840 0.0878200 0.5272
#> 370 0.055770 0.02281 0.0281500 0.1075
#> 371 0.055770 0.02281 0.0281500 0.1075
#> 372 0.040270 0.02901 0.0120600 0.1260
#> 373 0.072810 0.03070 0.0358300 0.1423
#> 374 0.052840 0.02348 0.0251400 0.1077
#> 375 0.072810 0.03070 0.0358300 0.1423
#> 376 0.031390 0.02493 0.0083390 0.1110
#> 377 0.069050 0.02718 0.0356800 0.1295
#> 378 0.145200 0.05540 0.0753300 0.2614
#> 379 0.055770 0.02281 0.0281500 0.1075
#> 380 0.119100 0.03250 0.0751200 0.1837
#> 381 0.052840 0.02348 0.0251400 0.1077
#> 382 0.050060 0.02446 0.0221200 0.1094
#> 383 0.107600 0.03283 0.0643400 0.1747
#> 384 0.020130 0.02076 0.0036240 0.1040
#> 385 0.019610 0.03257 0.0012320 0.2450
#> 386 0.076750 0.03521 0.0354100 0.1584
#> 387 0.076750 0.03521 0.0354100 0.1584
#> 388 0.138200 0.04695 0.0773800 0.2347
#> 389 0.029700 0.02385 0.0077860 0.1067
#> 390 0.131600 0.04015 0.0783400 0.2126
#> 391 0.131600 0.04015 0.0783400 0.2126
#> 392 0.058840 0.02265 0.0309200 0.1091
#> 393 0.055770 0.02281 0.0281500 0.1075
#> 394 0.025150 0.02184 0.0059250 0.1005
#> 395 0.119100 0.03250 0.0751200 0.1837
#> 396 0.055770 0.02281 0.0281500 0.1075
#> 397 0.023790 0.02146 0.0053020 0.1003
#> 398 0.052840 0.02348 0.0251400 0.1077
#> 399 0.052840 0.02348 0.0251400 0.1077
#> 400 0.044910 0.02678 0.0165600 0.1161
#> 401 0.042530 0.02793 0.0141700 0.1207
#> 402 0.042530 0.02793 0.0141700 0.1207
#> 403 0.087630 0.04404 0.0373700 0.1920
#> 404 0.040270 0.02901 0.0120600 0.1260
#> 405 0.074900 0.05232 0.0228500 0.2190
#> 406 0.025910 0.03330 0.0030270 0.1890
#> 407 0.080890 0.04066 0.0345600 0.1779
#> 408 0.031390 0.02493 0.0083390 0.1110
#> 409 0.023790 0.02146 0.0053020 0.1003
#> 410 0.069050 0.02718 0.0356800 0.1295
#> 411 0.065480 0.02470 0.0348200 0.1198
#> 412 0.023790 0.02146 0.0053020 0.1003
#> 413 0.052840 0.02348 0.0251400 0.1077
#> 414 0.038120 0.02998 0.0102300 0.1320
#> 415 0.038120 0.02998 0.0102300 0.1320
#> 416 0.036090 0.03082 0.0086410 0.1385
#> 417 0.131600 0.04015 0.0783400 0.2126
#> 418 0.107600 0.03283 0.0643400 0.1747
#> 419 0.044910 0.02678 0.0165600 0.1161
#> 420 0.067370 0.05650 0.0161900 0.2407
#> 421 0.052840 0.02348 0.0251400 0.1077
#> 422 0.092300 0.04092 0.0435500 0.1851
#> 423 0.080890 0.04066 0.0345600 0.1779
#> 424 0.076750 0.03521 0.0354100 0.1584
#> 425 0.076750 0.03521 0.0354100 0.1584
#> 426 0.131600 0.04015 0.0783400 0.2126
#> 427 0.040270 0.02901 0.0120600 0.1260
#> 428 0.071040 0.05455 0.0192600 0.2295