Summary method for gtReg

Produce a summary list for objects of class "gtReg" returned by gtReg.

Usage

# S3 method for class 'gtReg'
summary(object, ...)

Arguments

object

a fitted object of class "gtReg".

...

currently not used.

Value

summary.gtReg returns an object of class "summary.gtReg", a list containing:

call

the component from object.

link

the component from object.

deviance

the component from object, for simple pooling (type = "sp" in gtReg) only.

aic

the component from object, for simple pooling (type = "sp" in gtReg) only.

df.residual

the component from object, for simple pooling (type = "sp" in gtReg) only.

null.deviance

the component from object, for simple pooling (type = "sp" in gtReg) only.

df.null

the component from object, for simple pooling (type = "sp" in gtReg) only.

deviance.resid

the deviance residuals, for simple pooling (type = "sp" in gtReg) only.

coefficients

the matrix of coefficients, standard errors, z-values, and p-values. Aliased coefficients are omitted.

counts

the component from object.

method

the component from object, for simple pooling (type = "sp" in gtReg) only.

Gibbs.sample.size

the component from object, for array testing (type = "array" in gtReg) only.

cov.mat

the estimated covariance matrix of the estimated coefficients.

Details

The coefficients component of the results gives a matrix containing the estimated coefficients and their estimated standard errors. The third column is their ratio, labeled z ratio using Wald tests. A fourth column gives the two-tailed p-value corresponding to the z-ratio based on a Wald test. Note that it is possible that there are no residual degrees of freedom from which to estimate, in which case the estimate is NaN.

See also

Author

The majority of this function was originally written as summary.gt and summary.gt.mp by Boan Zhang for the binGroup package. Minor modifications were made to the function for inclusion in the binGroup2 package.

Examples

data(hivsurv)
fit1 <- gtReg(type = "sp",
              formula = groupres ~ AGE + EDUC.,
              data = hivsurv, groupn = gnum,
              sens = 0.9, spec = 0.9,
              method = "Xie")
#> 
#>  Number of minutes running: 0 
#>  
summary(fit1)
#> 
#> Call:
#> gtReg(type = "sp", formula = groupres ~ AGE + EDUC., data = hivsurv, 
#>     groupn = gnum, sens = 0.9, spec = 0.9, method = "Xie")
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -1.1729  -0.9406  -0.8281   1.3386   1.6497  
#> 
#> Coefficients:
#>             Estimate Std. Error z value Pr(>|z|)  
#> (Intercept) -3.11976    1.84816  -1.688   0.0914 .
#> AGE         -0.05692    0.07777  -0.732   0.4642  
#> EDUC.        0.82833    0.50717   1.633   0.1024  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#>     Null deviance: 154.3  on 85  degrees of freedom
#> Residual deviance: 109.5  on 83  degrees of freedom
#> AIC: 115.5
#> 
#> Number of iterations in EM: 43
#> 

# 5x6 and 4x5 array
set.seed(9128)
sa2a <- gtSim(type = "array", par = c(-7, 0.1),
              size1 = c(5, 4), size2 = c(6, 5),
              sens = 0.95, spec = 0.95)
sa2 <- sa2a$dframe
# \donttest{
fit2 <- gtReg(type = "array",
              formula = cbind(col.resp, row.resp) ~ x,
              data = sa2, coln = coln, rown = rown,
              arrayn = arrayn, sens = 0.95, spec = 0.95,
              linkf = "logit", n.gibbs = 1000, tol = 0.005)
#> 
#>  Number of minutes running: 0.07 
#>  
summary(fit2)# }
#> 
#> Call:
#> gtReg(type = "array", formula = cbind(col.resp, row.resp) ~ x, 
#>     data = sa2, coln = coln, rown = rown, arrayn = arrayn, sens = 0.95, 
#>     spec = 0.95, linkf = "logit", n.gibbs = 1000, tol = 0.005)
#> 
#> 
#> Coefficients:
#>             Estimate Std. Error z value Pr(>|z|)  
#> (Intercept) -6.24306    2.86818  -2.177   0.0295 *
#> x            0.08670    0.05982   1.450   0.1472  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Number of Gibbs samples generated in each E step: 1000
#> Number of iterations in EM algorithm: 3
#>