Extract the individual probabilities used to calculate group testing results

Extract the individual probabilities from objects of class "opchar" returned by operatingCharacteristics1 (opChar1) or operatingCharacteristics2 (opChar2).

IndProb(object, ...)

Arguments

object

An object of class "opChar", from which the individual probabilities are to be extracted.

...

Additional arguments to be passed to IndProb (e.g., digits to be passed to signif for appropriate rounding).

Value

Either p.vec, the sorted vector of individual probabilities (for hierarchical group testing algorithms) or p.mat, the sorted matrix of individual probabilities in gradient arrangement (for array testing algorithms). Further details are given under the 'Details' section for the operatingCharacteristics1 (opChar1) or operatingCharacteristics2 (opChar2) functions.

Author

Brianna D. Hitt

Examples

config.mat <- matrix(data = c(rep(1, 10), 1:10),
                     nrow = 2, ncol = 10, byrow = TRUE)
res1 <- opChar1(algorithm = "D2", p = 0.05, Se = 0.99, Sp = 0.99,
        hier.config = config.mat)
#> 
#>  Number of minutes running:  0 
#>  
IndProb(res1)
#>  [1] 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05

config.mat <- matrix(data = c(rep(1, 20), rep(1, 10), rep(2, 10),
                             rep(c(1, 2, 3, 4), each = 5),
                             rep(1, 3), rep(2, 2), rep(3, 3),
                             rep(4, 2), rep(5, 3), rep(6, 2),
                             rep(7, 3), rep(8, 2), 1:20),
                    nrow = 5, ncol = 20, byrow = TRUE)
Se <- matrix(data = rep(0.95, 10), nrow = 2, ncol = 5,
             dimnames = list(Infection = 1:2, Stage = 1:5))
Sp <- matrix(data = rep(0.99, 10), nrow = 2, ncol = 5,
             dimnames = list(Infection = 1:2, Stage = 1:5))
res2 <- opChar2(algorithm = "ID5",
                alpha = c(18.25, 0.75, 0.75, 0.25),
                Se = Se, Sp = Sp, hier.config = config.mat)
#> 
#>  Number of minutes running:  0 
#>  
IndProb(res2)
#>            1        2         3        4         5        6         7        8
#> 00 9.982e-01 0.977100 9.763e-01 0.971400 9.696e-01 0.961300 9.509e-01 0.945400
#> 10 2.892e-04 0.009650 6.683e-03 0.001454 4.554e-04 0.025980 2.133e-02 0.025670
#> 01 1.547e-03 0.010090 1.698e-02 0.024920 2.995e-02 0.005693 2.771e-02 0.023380
#> 11 1.003e-05 0.003159 2.635e-11 0.002201 3.267e-05 0.007018 6.705e-05 0.005502
#>           9        10        11        12       13        14      15       16
#> 00 0.941700 9.212e-01 9.174e-01 0.9163000 0.910300 9.098e-01 0.89830 0.896600
#> 10 0.004973 1.338e-02 7.834e-02 0.0008178 0.066940 8.165e-02 0.04963 0.089320
#> 01 0.007678 6.542e-02 5.196e-05 0.0765700 0.008664 8.538e-03 0.03591 0.002644
#> 11 0.045640 1.371e-06 4.245e-03 0.0063460 0.014130 1.174e-05 0.01617 0.011400
#>         17      18        19       20
#> 00 0.87560 0.83580 8.172e-01 0.778000
#> 10 0.06603 0.10180 6.164e-02 0.099180
#> 01 0.01651 0.02487 1.211e-01 0.120100
#> 11 0.04187 0.03753 1.076e-12 0.002681