Summary measures for the Thresholded Optimal Dorfman (TOD) algorithm

Summary measures for the Thresholded Optimal Dorfman (TOD) algorithm.

TOD(p.vec, Se, Sp, max = 15, init.group.sz = NULL, threshold = NULL)

Arguments

p.vec

a vector of individual risk probabilities.

Se

sensitivity of the diagnostic test.

Sp

specificity of the diagnostic test.

max

the maximum allowable group size. Further details are given under 'Details'.

init.group.sz

the initial group size used for TOD, if threshold is not specified. Further details are given under 'Details'.

threshold

the threshold value for TOD. If a threshold is not specified, one is found algorithmically. Further details are given under 'Details'.

Value

A list containing:

prob

the vector of individual risk probabilities, as specified by the user.

Se

the sensitivity of the diagnostic test, as specified by the user.

Sp

the specificity of the diagnostic test, as specified by the user.

group.sz

the initial group size used for TOD, if applicable.

thresh.val

the threshold value used for TOD, if applicable.

OTC

a list specifying elements of the optimal testing configuration, which may include:

Block.sz

the block size/initial group size for informative Dorfman testing, which is not tested.

pool.szs

group sizes for the first stage of testing for informative Dorfman testing.

ET

the expected testing expenditure to decode all individuals in the algorithm.

Var

the variance of the testing expenditure to decode all individuals in the algorithm.

Accuracy

a list containing:

Individual

a matrix of accuracy measures for each individual. The rows correspond to each unique set of accuracy measures in the algorithm. Individuals with the same set of accuracy measures are displayed together in a single row of the matrix. The columns correspond to the pool index, the individual risk probability, and the pooling sensitivity, pooling specificity, pooling positive predictive value, and pooling negative predictive value for the individuals in each row of the matrix.

Overall

a matrix of overall accuracy measures for the algorithm. The columns correspond to the pooling sensitivity, pooling specificity, pooling positive predictive value, and pooling negative predictive value for the overall algorithm. Further details are given under 'Details'.

Details

This function finds the characteristics of an informative two-stage hierarchical (Dorfman) decoding process. Characteristics found include the expected expenditure of the decoding process, the variance of the expenditure of the decoding process, and the pooling sensitivity, pooling specificity, pooling positive predictive value, and pooling negative predictive value for each individual and for the overall algorithm. Calculations of these characteristics are done using equations presented in McMahan et al. (2012).

Thresholded Optimal Dorfman (TOD) is an informative Dorfman algorithm in which all \(N\) individuals are partitioned into two classes, low-risk and high-risk individuals. The threshold can be specified using the optional threshold argument. Alternatively, the TOD algorithm can identify the optimal threshold value. The low-risk individuals are tested using an optimal common pool size, and the high-risk individuals are tested individually. If desired, the user can add the constraint of a maximum allowable group size (max), so that each group will contain no more than the maximum allowable number of individuals.

The displayed overall pooling sensitivity, pooling specificity, pooling positive predictive value, and pooling negative predictive value are weighted averages of the corresponding individual accuracy measures for all individuals within the initial group (or block) for a hierarchical algorithm, or within the entire array for an array-based algorithm. Expressions for these averages are provided in the Supplementary Material for Hitt et al. (2019). These expressions are based on accuracy definitions given by Altman and Bland (1994a, 1994b).

References

Altman, D., Bland, J. (1994). “Diagnostic tests 1: Sensitivity and specificity.” BMJ, 308, 1552.

Altman, D., Bland, J. (1994). “Diagnostic tests 2: Predictive values.” BMJ, 309, 102.

Hitt, B., Bilder, C., Tebbs, J., McMahan, C. (2019). “The objective function controversy for group testing: Much ado about nothing?” Statistics in Medicine, 38, 4912–4923.

McMahan, C., Tebbs, J., Bilder, C. (2012a). “Informative Dorfman Screening.” Biometrics, 68, 287–296.

See also

Author

Brianna D. Hitt

Examples

# Example 1: Find the characteristics of an informative
#   Dorfman algorithm, using the TOD procedure.
set.seed(1002)
p.vec <- expectOrderBeta(p = 0.01, alpha = 2, size = 20)
TOD(p = p.vec, Se = 0.95, Sp = 0.95, max = 5,
    threshold = 0.015)
#> Overall block size: 20 
#> Ordered individual probabilities of positivity: 
#>  0.0016 0.0025 0.0033 0.004 0.0047 0.0053 0.006 0.0067 0.0074 0.0082 0.0089 0.0098 0.0107 0.0117 0.0129 0.0143 0.0159 0.018 0.0211 0.027 
#> Group sizes for individuals ordered by probabilities of positivity: 
#>  4 4 4 4 1 1 1 1 
#>  
#> Expected number of tests: 9.22 
#> Variance for the number of tests: 4.49 

# Example 2: Find the threshold value for the TOD
#   procedure algorithmically. Then, find
#   characteristics of the algorithm.
TOD(p = p.vec, Se = 0.95, Sp = 0.95, max = 5,
    init.group.sz = 10)
#> Overall block size: 20 
#> Ordered individual probabilities of positivity: 
#>  0.0016 0.0025 0.0033 0.004 0.0047 0.0053 0.006 0.0067 0.0074 0.0082 0.0089 0.0098 0.0107 0.0117 0.0129 0.0143 0.0159 0.018 0.0211 0.027 
#> Group sizes for individuals ordered by probabilities of positivity: 
#>  5 5 5 5 
#>  
#> Expected number of tests: 5.88 
#> Variance for the number of tests: 8.43