Determine a vector of probabilities for informative group testing algorithms

Find the expected value of order statistics from a beta distribution. This function is used to provide a set of individual risk probabilities for informative group testing.

expectOrderBeta(
  p,
  alpha,
  size,
  grp.sz,
  num.sim = 10000,
  rel.tol = ifelse(alpha >= 1, .Machine$double.eps^0.25, .Machine$double.eps^0.1),
  ...
)

Arguments

p

overall probability of disease that will be used to determine a vector of individual risk probabilities. This is the expected value of a random variable with a beta distribution, \(\frac{\alpha}{\alpha + \beta}\).

alpha

a shape parameter for the beta distribution that specifies the degree of heterogeneity for the determined probability vector.

size

the size of the vector of individual risk probabilities to be generated. This is also the number of total individuals for which to determine risk probabilities.

grp.sz

the number of total individuals for which to determine risk probabilities. This argument is deprecated; the size argument should be used instead.

num.sim

the number of simulations. This argument is used only when simulation is necessary.

rel.tol

relative tolerance used for integration.

...

arguments to be passed to the beta.dist function written by Michael Black for Black et al. (2015).

Value

A vector of individual risk probabilities.

Details

This function uses the beta.dist function from Black et al. (2015) to determine a vector of individual risk probabilities, ordered from least to greatest. Depending on the specified probability, \(\alpha\) level, and overall group size, simulation may be necessary in order to determine the probabilities. For this reason, the user should set a seed in order to reproduce results. The number of simulations (default = 10,000) and relative tolerance for integration can be specified by the user. The expectOrderBeta function augments the beta.dist function by checking whether simulation is needed before attempting to determine the probabilities, and by allowing the number of simulations to be specified by the user. See Black et al. (2015) for additional details on the original beta.dist function.

References

Black, M., Bilder, C., Tebbs, J. (2015). “Optimal retesting configurations for hierarchical group testing.” Journal of the Royal Statistical Society. Series C: Applied Statistics, 64, 693–710.

See also

Author

Brianna D. Hitt

Examples

set.seed(8791)
expectOrderBeta(p = 0.03, alpha = 0.5, size = 100, rel.tol = 0.0001)
#> Using simulation
#>   [1] 9.888742e-06 2.960985e-05 5.758418e-05 9.536279e-05 1.435112e-04
#>   [6] 2.014517e-04 2.693015e-04 3.484398e-04 4.347717e-04 5.301154e-04
#>  [11] 6.347749e-04 7.553283e-04 8.813811e-04 1.019173e-03 1.165680e-03
#>  [16] 1.321324e-03 1.487075e-03 1.665317e-03 1.853131e-03 2.052074e-03
#>  [21] 2.263881e-03 2.484335e-03 2.716080e-03 2.960775e-03 3.216792e-03
#>  [26] 3.485977e-03 3.768618e-03 4.055657e-03 4.361395e-03 4.674162e-03
#>  [31] 5.003723e-03 5.343320e-03 5.697125e-03 6.066926e-03 6.444940e-03
#>  [36] 6.839459e-03 7.252745e-03 7.681719e-03 8.121652e-03 8.586982e-03
#>  [41] 9.065121e-03 9.554287e-03 1.005516e-02 1.058576e-02 1.112961e-02
#>  [46] 1.169173e-02 1.227191e-02 1.287684e-02 1.349762e-02 1.414197e-02
#>  [51] 1.479939e-02 1.547656e-02 1.616789e-02 1.689998e-02 1.765705e-02
#>  [56] 1.845073e-02 1.925670e-02 2.006631e-02 2.093889e-02 2.183414e-02
#>  [61] 2.276064e-02 2.369815e-02 2.469039e-02 2.572719e-02 2.677131e-02
#>  [66] 2.788736e-02 2.904225e-02 3.024986e-02 3.149846e-02 3.281283e-02
#>  [71] 3.417198e-02 3.556576e-02 3.704478e-02 3.857573e-02 4.017339e-02
#>  [76] 4.183985e-02 4.360637e-02 4.546013e-02 4.739956e-02 4.945612e-02
#>  [81] 5.161278e-02 5.390850e-02 5.633163e-02 5.889587e-02 6.166592e-02
#>  [86] 6.462436e-02 6.784261e-02 7.127314e-02 7.500320e-02 7.918243e-02
#>  [91] 8.381123e-02 8.890606e-02 9.461215e-02 1.011622e-01 1.088779e-01
#>  [96] 1.181433e-01 1.297517e-01 1.454834e-01 1.679224e-01 2.120121e-01

expectOrderBeta(p = 0.05, alpha = 2, size = 40)
#>  [1] 0.005570409 0.008630251 0.011086767 0.013256762 0.015259102 0.017153566
#>  [7] 0.018975479 0.020748047 0.022487753 0.024207057 0.025915887 0.027622530
#> [13] 0.029334200 0.031057425 0.032798321 0.034562810 0.036356794 0.038186309
#> [19] 0.040057678 0.041977661 0.043953627 0.045993740 0.048107195 0.050304504
#> [25] 0.052597857 0.055001598 0.057532868 0.060212467 0.063066087 0.066126076
#> [31] 0.069434074 0.073045078 0.077034015 0.081506906 0.086621178 0.092625871
#> [37] 0.099951032 0.109442444 0.123165239 0.149033333