Auxiliary for controlling group testing regression

Auxiliary function to control fitting parameters of the EM algorithm used internally in gtReg for simple pooling (type = "sp") with method = "Xie" or for array testing (type = "array").

gtRegControl(
  tol = 1e-04,
  n.gibbs = 1000,
  n.burnin = 20,
  maxit = 500,
  trace = FALSE,
  time = TRUE
)

Arguments

tol

convergence criterion.

n.gibbs

the Gibbs sample size to be used in each E step of the EM algorithm, for array testing. The default is 1000.

n.burnin

the number of samples in the burn-in period, for array testing. The default is 20.

maxit

maximum number of iterations in the EM algorithm.

trace

a logical value indicating whether the output should be printed for each iteration. The default is FALSE.

time

a logical value indicating whether the length of time for the model fitting should be printed. The default is TRUE.

Value

A list with components named as the input arguments.

Author

This function was originally written as the gt.control function for the binGroup package. Minor modifications have been made for inclusion in the binGroup2 package.

Examples

# The default settings:
gtRegControl()
#> $tol
#> [1] 1e-04
#> 
#> $n.gibbs
#> [1] 1000
#> 
#> $n.burnin
#> [1] 20
#> 
#> $maxit
#> [1] 500
#> 
#> $trace
#> [1] FALSE
#> 
#> $time
#> [1] TRUE
#>