Simulation function for group testing data

Simulates data in group testing form ready to be fit by gtReg.

gtSim(
  type = "sp",
  x = NULL,
  gshape = 20,
  gscale = 2,
  par,
  linkf = c("logit", "probit", "cloglog"),
  size1,
  size2,
  sens = 1,
  spec = 1,
  sens.ind = NULL,
  spec.ind = NULL
)

Arguments

type

"sp" for simple pooling (Dorfman testing with or without retests), "halving" for halving protocol, and "array" for array testing (also known as matrix pooling).

x

a matrix of user-submitted covariates with which to simulate the data. Default is NULL, in which case a gamma distribution is used to generate the covariates automatically.

gshape

shape parameter for the gamma distribution. The value must be non-negative. Default value is set to 20.

gscale

scale parameter for the gamma distribution. The value must be strictly positive. Default value is set to 2.

par

the true coefficients in the linear predictor.

linkf

a character string specifying one of the three link functions to be used: "logit" (default), "probit", or "cloglog".

size1

sample size of the simulated data (for use with "sp" and "halving" methods) or a vector that specifies the number of rows in each matrix (for use with "array" method). If only one matrix is simulated, this value is a scalar.

size2

group size in pooling individual samples (for use with "sp" and "halving" methods) or a vector that specifies the number of columns in each matrix (for use with "array" method). If only one matrix is simulated, this value is a scalar.

sens

sensitivity of the group tests. Default value is set to 1.

spec

specificity of the group tests. Default value is set to 1.

sens.ind

sensitivity of the individual retests. If NULL, set to be equal to sens.

spec.ind

specificity of the individual retests. If NULL, set to be equal to spec.

Value

For simple pooling (type = "sp") and the halving protocol (type = "halving"), a data frame or for array testing (type = "array"), a list, which may include the following:

gres

the group response, for simple pooling and the halving protocol only.

col.resp

the column group response, for array testing only.

row.resp

the row group response, for array testing only.

x

the covariate.

groupn

the group number, for simple pooling and the halving protocol only.

arrayn

the array number, for array testing only.

coln

the column group number, for array testing only.

rown

the row group number, for array testing only.

ind

the true individual responses. For simple pooling and the halving protocol, these are included in the data frame of results. For array testing, these are included in the list of results, with individual responses presented in matrices.

retest

the results of individual retests.

subgroup

the subgroup number, for the halving protocol.

prob

the individual probabilities, for array testing only.

Details

Generates group testing data in simple pooling form (type = "sp"), for the halving protocol (type = "halving"), or in array testing form (type = "array"). The covariates are either specified by the x argument or they are generated from a gamma distribution with the given gshape and gscale parameters. The individual probabilities are calculated from the covariates, the coefficients given in par, and the link function specified through linkf. The true binary individual responses are then simulated from the individual probabilities.

Under the matrix pooling protocol (type = "array"), the individuals are first organized into (by column) one or more matrices specified by the number of rows (size1) and the number of columns (size2).

Then, for all pooling protocols, the true group responses are found from the individual responses within groups or within rows/columns for matrix pooling (i.e., if at least one response is positive, the group is positive; otherwise, the group response is negative). Finally, the observed group (method = "sp") and subgroup method = "halving" only), or row and column responses method = "array" are simulated using the given sens and spec.

For the simple pooling and halving protocols, individual retests are simulated from sens.ind and spec.ind for samples in observed positive groups. Note that with a given group size (specified by size2 with method = "sp" or method = "halving"), the last group may have fewer individuals. For the matrix pooling protocol, individual retests are simulated from sens.ind and spec.ind for individuals that lie on the intersection of an observed positive row and and observed positive column. In the case where no column (row) tests positive in a matrix, all the individuals in any observed positive rows (columns) will be assigned a simulated retest result. If no column or row is observed positive, NULL is returned.

See also

Author

This function is a combination of sim.gt, sim.halving, and sim.mp written by Boan Zhang for the binGroup package. Minor modifications have been made for inclusion of the functions in the binGroup2 package.

Examples

set.seed(46)
gt.data <- gtSim(type = "sp", par = c(-12, 0.2),
                 size1 = 700, size2 = 5)

x1 <- sort(runif(100, 0, 30))
x2 <- rgamma(100, shape = 17, scale = 1.5)
gt.data <- gtSim(type = "sp", x = cbind(x1, x2),
                 par = c(-14, 0.2, 0.3), size2 = 4,
                 sens = 0.98, spec = 0.98)

set.seed(46)
gt.data <- gtSim(type = "halving", par = c(-6, 0.1),
                 gshape = 17, gscale = 1.4, size1 = 5000,
                 size2 = 5, sens = 0.95, spec = 0.95)

# 5x6 and 4x5 matrix
set.seed(9128)
sa1a <- gtSim(type = "array", par = c(-7, 0.1),
              size1 = c(5, 4), size2 = c(6, 5),
              sens = 0.95, spec = 0.95)
sa1a$dframe
#>           x col.resp row.resp coln rown arrayn retest
#> 1  29.96059        0        0    1    1      1     NA
#> 2  61.28205        0        1    1    2      1     NA
#> 3  34.27341        0        1    1    3      1     NA
#> 4  46.19001        0        0    1    4      1     NA
#> 5  39.43801        0        1    1    5      1     NA
#> 6  45.88038        1        0    2    1      1     NA
#> 7  37.46931        1        1    2    2      1      0
#> 8  41.53810        1        1    2    3      1      0
#> 9  33.19917        1        0    2    4      1     NA
#> 10 39.12554        1        1    2    5      1      1
#> 11 35.63487        1        0    3    1      1     NA
#> 12 76.53439        1        1    3    2      1      1
#> 13 43.28924        1        1    3    3      1      0
#> 14 19.54699        1        0    3    4      1     NA
#> 15 42.47747        1        1    3    5      1      0
#> 16 50.58869        0        0    4    1      1     NA
#> 17 26.54039        0        1    4    2      1     NA
#> 18 42.56257        0        1    4    3      1     NA
#> 19 53.56499        0        0    4    4      1     NA
#> 20 37.74931        0        1    4    5      1     NA
#> 21 41.87551        0        0    5    1      1     NA
#> 22 40.20522        0        1    5    2      1     NA
#> 23 37.97127        0        1    5    3      1     NA
#> 24 38.45669        0        0    5    4      1     NA
#> 25 39.56397        0        1    5    5      1     NA
#> 26 37.93089        1        0    6    1      1     NA
#> 27 42.63813        1        1    6    2      1      1
#> 28 35.52633        1        1    6    3      1      0
#> 29 37.21599        1        0    6    4      1     NA
#> 30 35.73672        1        1    6    5      1      0
#> 31 37.80483        0        0    1    1      2     NA
#> 32 45.86955        0        0    1    2      2     NA
#> 33 46.37589        0        1    1    3      2      1
#> 34 33.38453        0        0    1    4      2     NA
#> 35 30.46027        0        0    2    1      2     NA
#> 36 42.76132        0        0    2    2      2     NA
#> 37 32.44970        0        1    2    3      2      0
#> 38 36.08952        0        0    2    4      2     NA
#> 39 41.96027        0        0    3    1      2     NA
#> 40 45.06366        0        0    3    2      2     NA
#> 41 51.31051        0        1    3    3      2      0
#> 42 34.06011        0        0    3    4      2     NA
#> 43 50.30978        0        0    4    1      2     NA
#> 44 32.58809        0        0    4    2      2     NA
#> 45 26.49616        0        1    4    3      2      0
#> 46 27.90731        0        0    4    4      2     NA
#> 47 30.96149        0        0    5    1      2     NA
#> 48 33.31087        0        0    5    2      2     NA
#> 49 44.68050        0        1    5    3      2      0
#> 50 42.13516        0        0    5    4      2     NA