Probability mass function for halving

Calculate the probability mass function for the number of tests from using the halving algorithm.

halving(p, Se = 1, Sp = 1, stages = 2, order.p = TRUE)

Arguments

p

a vector of individual risk probabilities.

Se

sensitivity of the diagnostic test.

Sp

specificity of the diagnostic test.

stages

the number of stages for the halving algorithm.

order.p

logical; if TRUE, the vector of individual risk probabilities will be sorted.

Value

A list containing:

pmf

the probability mass function for the halving algorithm.

et

the expected testing expenditure for the halving algorithm.

vt

the variance of the testing expenditure for the halving algorithm.

p

a vector containing the probabilities of positivity for each individual.

Details

Halving algorithms involve successively splitting a positive testing group into two equal-sized halves (or as close to equal as possible) until all individuals have been identified as positive or negative. \(S\)-stage halving begins by testing the whole group of \(I\) individuals. Positive groups are split in half until the final stage of the algorithm, which consists of individual testing. For example, consider an initial group of size \(I=16\) individuals. Three-stage halving (3H) begins by testing the whole group of 16 individuals. If this group tests positive, the second stage involves splitting into two groups of size 8. If either of these groups test positive, a third stage involves testing each individual rather than halving again. Four-stage halving (4H) would continue with halving into groups of size 4 before individual testing. Five-stage halving (5H) would continue with halving into groups of size 2 before individual testing. 3H requires more than 2 individuals, 4H requires more than 4 individuals, and 5H requires more than 8 individuals.

This function calculates the probability mass function, expected testing expenditure, and variance of the testing expenditure for halving algorithms with 3 to 5 stages.

References

Black, M., Bilder, C., Tebbs, J. (2012). “Group testing in heterogeneous populations by using halving algorithms.” Journal of the Royal Statistical Society. Series C: Applied Statistics, 61, 277–290.

See also

Author

This function was originally written by Michael Black for Black et al. (2012). The function was obtained from http://chrisbilder.com/grouptesting/. Minor modifications have been made for inclusion of the function in the binGroup2 package.

Examples

# Equivalent to Dorfman testing (two-stage hierarchical)
halving(p = rep(0.01, 10), Se = 1, Sp = 1, stages = 2,
        order.p = TRUE)
#> pmf :
#>   num.tests prob.tests
#> 1         1     0.9044
#> 2        11     0.0956
#> 
#> et :
#> [1] 1.956179
#> 
#> vt :
#> [1] 8.647514
#> 
#> p :
#>  [1] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> 

# Halving over three stages; each individual has a
#   different probability of being positive
set.seed(12895)
p.vec <- expectOrderBeta(p = 0.05, alpha = 2, size = 20)
halving(p = p.vec, Se = 0.95, Sp = 0.95, stages = 3,
        order.p = TRUE)
#> pmf :
#>   num.tests prob.tests
#> 1         1     0.3690
#> 2         3     0.0399
#> 3        13     0.4624
#> 4        23     0.1287
#> 
#> et :
#> [1] 9.459485
#> 
#> vt :
#> [1] 57.46018
#> 
#> p :
#>  [1] 0.00813011 0.01281577 0.01672941 0.02032010 0.02376267 0.02715206
#>  [7] 0.03055215 0.03401377 0.03758344 0.04130894 0.04524401 0.04945382
#> [13] 0.05402248 0.05906516 0.06474940 0.07133702 0.07927673 0.08944686
#> [19] 0.10398454 0.13105154
#>