Data from an HIV surveillance project

The hivsurv data set comes from an HIV surveillance project discussed in Verstraeten et al. (1998) and Vansteelandt et al. (2000). The purpose of the study was to estimate the HIV prevalence among pregnant Kenyan women in four rural locations of the country, using both individual and group testing responses. Blood tests were administered to each participating woman, and 4 covariates were obtained on each woman. Because the original group responses are unavailable, individuals are artificially put into groups of 5 here to form group responses. Only the 428 complete observations are given.

data(hivsurv)

Format

A data frame with 428 observations on the following 8 variables.

DATE

the date when each sample was collected.

PAR.

parity (number of children).

AGE

age (in years).

MA.ST.

marital status (1: single; 2: married (polygamous); 3: married (monogamous); 4: divorced; 5: widow).

EDUC.

highest attained education level (1: no schooling; 2: primary school; 3: secondary school; 4: higher).

HIV

individual response of HIV diagnosis (0: negative; 1: positive).

gnum

the group number that designates individuals into groups.

groupres

the group response calculated from artificially formed groups.

Source

Vansteelandt, S., Goetghebeur, E., Verstraeten, T. (2000). “Regression models for disease prevalence with diagnostic tests on pools of serum samples.” Biometrics, 56, 1126--1133.

Verstraeten, T., Farah, B., Duchateau, L., Matu, R. (1998). “Pooling sera to reduce the cost of HIV surveillance: a feasibility study in a rural Kenyan district.” Tropical Medicine & International Health, 3, 747--750.

Examples

data(hivsurv)

str(hivsurv)
#> 'data.frame':	428 obs. of  8 variables:
#>  $ DATE    : Factor w/ 64 levels "10/1/1996","10/11/1996",..: 47 53 53 53 53 53 53 53 60 60 ...
#>  $ PAR.    : int  0 1 0 1 1 3 2 5 0 0 ...
#>  $ AGE     : int  21 16 17 17 18 20 26 35 17 18 ...
#>  $ MA.ST.  : int  3 3 3 3 3 3 3 3 3 3 ...
#>  $ EDUC.   : int  4 2 1 2 1 2 2 2 2 2 ...
#>  $ HIV     : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ gnum    : int  1 1 1 1 1 2 2 2 2 2 ...
#>  $ groupres: int  0 0 0 0 0 0 0 0 0 0 ...