Summary method for operating characteristics results

Produce a summary list for objects of class "opChar" returned by operatingCharacteristics1 (opChar1) or operatingCharacteristics2 (opChar2).

# S3 method for opChar
summary(object, ...)

Arguments

object

an object of class "opChar", providing the calculated operating characteristics for a group testing algorithm.

...

currently not used.

Value

summary.opChar returns an object of class "summary.opChar", a list containing:

Algorithm

character string specifying the name of the group testing algorithm.

Configuration

matrix detailing the configuration from object. For hierarchical testing, this includes the group sizes for each stage of testing. For array testing, this includes the array dimension (row/column size) and the array size (the total number of individuals in the array).

Tests

matrix detailing the expected number of tests and expected number of tests per individual from object

.

Accuracy

a list containing:

Individual

matrix detailing the accuracy measures for each individual from object (for objects returned by opChar1).

Disease 1 Individual

matrix detailing the accuracy measures pertaining to disease 1 for each individual from object (for objects returned by opChar2).

Disease 2 Individual

matrix detailing the accuracy measures pertaining to disease 2 for each individual from object (for objects returned by opChar2).

Overall

matrix detailing the overall accuracy measures for the algorithm from object.

Details

This function produces a summary list for objects of class "opChar" returned by operatingCharacteristics1 (opChar1) or operatingCharacteristics2 (opChar2). It formats the testing configuration, expected number of tests, expected number of tests per individual, and accuracy measures.

The Configuration component of the result gives the testing configuration, which may include the group sizes for each stage of a hierarchical testing algorithm or the row/column size and array size for an array testing algorithm. The Tests component of the result gives the expected number of tests and the expected number of tests per individual for the algorithm.

The Accuracy component gives the individual accuracy measures for each individual in object and the overall accuracy measures for the algorithm. Accuracy measures included are the pooling sensitivity, pooling specificity, pooling positive predictive value, and pooling negative predictive value. The overall accuracy measures displayed are weighted averages of the corresponding individual accuracy measures for all individuals in the algorithm. Expressions for these averages are provided in the Supplementary Material for Hitt et al. (2019). For more information, see the Details' section for the operatingCharacteristics1 (opChar1) or operatingCharacteristics2 (opChar2) function.

See also

Author

Brianna D. Hitt

Examples

# Calculate the operating characteristics for
#   non-informative four-stage hierarchical testing.
config.mat <- matrix(data = c(rep(1, 24), rep(1, 16),
                              rep(2, 8), rep(1, 8),
                              rep(2, 8), rep(3, 4),
                              rep(4, 2), rep(5, 2), 1:24),
                     nrow = 4, ncol = 24, byrow = TRUE)
calc1 <- opChar1(algorithm = "D4", p = 0.01,
                 Se = 0.99, Sp = 0.99,
                 hier.config = config.mat,
                 a = c(1, 9, 17, 21, 23))
#> 
#>  Number of minutes running:  0 
#>  
summary(calc1)
#> 
#> Algorithm: Non-informative four-stage hierarchical testing 
#> 
#> Testing configuration:
#> Stage 1: 24
#> Stage 2: 16,8
#> Stage 3: 8,8,4,2,2
#> 
#> Expected number of tests: 3.41
#> Expected number of tests per individual: 0.1420
#> 
#> Accuracy for individuals:
#>      PSe    PSp   PPPV   PNPV Individuals
#> 1 0.9606 0.9993 0.9357 0.9996         1,9
#> 2 0.9606 0.9997 0.9708 0.9996          17
#> 3 0.9606 0.9999 0.9895 0.9996       21,23
#> 
#> Overall accuracy of the algorithm:
#>      PSe    PSp   PPPV   PNPV
#> 1 0.9606 0.9995 0.9501 0.9996
#> 
#> PSe denotes the pooling sensitivity.
#> PSp denotes the pooling specificity.
#> PPPV denotes the pooling positive predictive value.
#> PNPV denotes the pooling negative predictive value.

# Calculate the operating characteristics for
#   informative array testing without master pooling.
calc2 <- opChar1(algorithm = "IA2", p = 0.025, alpha = 0.5,
                 Se = 0.95, Sp = 0.99, rowcol.sz = 12)
#> 
#>  Number of minutes running:  0.01 
#>  
summary(calc2)
#> 
#> Algorithm: Informative array testing without master pooling 
#> 
#> Testing configuration:
#> Row/column size: 12
#> Array size: 144
#> 
#> Expected number of tests: 32.82
#> Expected number of tests per individual: 0.2279
#> 
#> Accuracy for individuals:
#>        PSe    PSp   PPPV   PNPV Individuals
#> 1   0.8650 0.9987 0.9936 0.9686           1
#> 2   0.8625 0.9986 0.9801 0.9892           2
#> 3   0.8619 0.9989 0.9751 0.9931           3
#> 4   0.8616 0.9992 0.9727 0.9952           4
#> 5   0.8613 0.9994 0.9677 0.9972           5
#> 6   0.8612 0.9996 0.9640 0.9981           6
#> 7   0.8612 0.9997 0.9645 0.9986           7
#> 8   0.8611 0.9998 0.9626 0.9992           8
#> 9   0.8610 0.9999 0.9466 0.9996           9
#> 10  0.8610 0.9999 0.9451 0.9997          10
#> 11  0.8610 0.9999 0.8685 0.9999          11
#> 12  0.8610 0.9999 0.3963 1.0000          12
#> 13  0.8649 0.9987 0.9918 0.9756          13
#> 14  0.8624 0.9988 0.9815 0.9897          14
#> 15  0.8618 0.9990 0.9770 0.9933          15
#> 16  0.8615 0.9992 0.9748 0.9954          16
#> 17  0.8612 0.9995 0.9700 0.9973          17
#> 18  0.8611 0.9996 0.9670 0.9981          18
#> 19  0.8610 0.9997 0.9667 0.9986          19
#> 20  0.8610 0.9998 0.9651 0.9992          20
#> 21  0.8609 0.9999 0.9513 0.9996          21
#> 22  0.8609 0.9999 0.9445 0.9998          22
#> 23  0.8609 0.9999 0.8658 0.9999          23
#> 24  0.8609 0.9999 0.2376 1.0000          24
#> 25  0.8649 0.9987 0.9907 0.9791          25
#> 26  0.8624 0.9989 0.9820 0.9901          26
#> 27  0.8618 0.9991 0.9778 0.9935          27
#> 28  0.8614 0.9993 0.9757 0.9955          28
#> 29  0.8612 0.9995 0.9711 0.9974          29
#> 30  0.8611 0.9996 0.9685 0.9982          30
#> 31  0.8610 0.9997 0.9683 0.9987          31
#> 32  0.8609 0.9998 0.9646 0.9993          32
#> 33  0.8609 0.9999 0.9528 0.9996          33
#> 34  0.8609 0.9999 0.9452 0.9998          34
#> 35  0.8609 0.9999 0.8354 1.0000          35
#> 36  0.8608 1.0000 0.1137 1.0000          36
#> 37  0.8648 0.9987 0.9898 0.9813          37
#> 38  0.8623 0.9989 0.9823 0.9905          38
#> 39  0.8617 0.9991 0.9783 0.9937          39
#> 40  0.8614 0.9993 0.9763 0.9956          40
#> 41  0.8611 0.9995 0.9717 0.9975          41
#> 42  0.8610 0.9996 0.9690 0.9982          42
#> 43  0.8610 0.9997 0.9699 0.9987          43
#> 44  0.8609 0.9998 0.9626 0.9994          44
#> 45  0.8608 0.9999 0.9539 0.9996          45
#> 46  0.8608 0.9999 0.9472 0.9998          46
#> 47  0.8608 0.9999 0.8373 1.0000          47
#> 48  0.8608 1.0000 0.0445 1.0000          48
#> 49  0.8648 0.9988 0.9891 0.9830          49
#> 50  0.8623 0.9990 0.9824 0.9908          50
#> 51  0.8617 0.9992 0.9786 0.9939          51
#> 52  0.8613 0.9994 0.9767 0.9957          52
#> 53  0.8611 0.9996 0.9721 0.9975          53
#> 54  0.8610 0.9997 0.9701 0.9982          54
#> 55  0.8609 0.9998 0.9694 0.9987          55
#> 56  0.8608 0.9999 0.9623 0.9994          56
#> 57  0.8608 0.9999 0.9527 0.9997          57
#> 58  0.8608 0.9999 0.9343 0.9998          58
#> 59  0.8608 1.0000 0.8328 1.0000          59
#> 60  0.8608 1.0000 0.0146 1.0000          60
#> 61  0.8648 0.9988 0.9884 0.9843          61
#> 62  0.8623 0.9990 0.9825 0.9912          62
#> 63  0.8617 0.9992 0.9788 0.9941          63
#> 64  0.8613 0.9994 0.9769 0.9959          64
#> 65  0.8611 0.9996 0.9723 0.9976          65
#> 66  0.8610 0.9997 0.9707 0.9983          66
#> 67  0.8609 0.9998 0.9690 0.9988          67
#> 68  0.8608 0.9999 0.9632 0.9994          68
#> 69  0.8608 0.9999 0.9533 0.9997          69
#> 70  0.8608 0.9999 0.9350 0.9998          70
#> 71  0.8608 1.0000 0.7795 1.0000          71
#> 72  0.8608 1.0000 0.0040 1.0000          72
#> 73  0.8647 0.9988 0.9879 0.9853          73
#> 74  0.8622 0.9990 0.9825 0.9915          74
#> 75  0.8616 0.9992 0.9790 0.9943          75
#> 76  0.8613 0.9994 0.9771 0.9960          76
#> 77  0.8610 0.9996 0.9725 0.9977          77
#> 78  0.8610 0.9997 0.9715 0.9983          78
#> 79  0.8609 0.9998 0.9688 0.9988          79
#> 80  0.8608 0.9999 0.9590 0.9995          80
#> 81  0.8608 0.9999 0.9544 0.9997          81
#> 82  0.8607 0.9999 0.9353 0.9998          82
#> 83  0.8607 1.0000 0.7798 1.0000          83
#> 84  0.8607 1.0000 0.0009 1.0000          84
#> 85  0.8647 0.9989 0.9875 0.9862          85
#> 86  0.8622 0.9991 0.9826 0.9918          86
#> 87  0.8616 0.9993 0.9792 0.9945          87
#> 88  0.8613 0.9994 0.9774 0.9961          88
#> 89  0.8610 0.9996 0.9728 0.9978          89
#> 90  0.8609 0.9997 0.9710 0.9984          90
#> 91  0.8609 0.9998 0.9648 0.9990          91
#> 92  0.8608 0.9999 0.9559 0.9995          92
#> 93  0.8607 0.9999 0.9526 0.9997          93
#> 94  0.8607 0.9999 0.9206 0.9999          94
#> 95  0.8607 1.0000 0.7752 1.0000          95
#> 96  0.8607 1.0000 0.0002 1.0000          96
#> 97  0.8647 0.9989 0.9871 0.9870          97
#> 98  0.8622 0.9991 0.9825 0.9921          98
#> 99  0.8616 0.9993 0.9793 0.9946          99
#> 100 0.8613 0.9995 0.9774 0.9962         100
#> 101 0.8610 0.9996 0.9728 0.9978         101
#> 102 0.8609 0.9997 0.9717 0.9984         102
#> 103 0.8608 0.9998 0.9654 0.9990         103
#> 104 0.8608 0.9999 0.9563 0.9995         104
#> 105 0.8607 0.9999 0.9527 0.9997         105
#> 106 0.8607 0.9999 0.9162 0.9999         106
#> 107 0.8607 1.0000 0.7306 1.0000         107
#> 108 0.8607 1.0000 0.0000 1.0000         108
#> 109 0.8647 0.9990 0.9870 0.9877         109
#> 110 0.8622 0.9992 0.9828 0.9924         110
#> 111 0.8616 0.9993 0.9797 0.9948         111
#> 112 0.8612 0.9995 0.9722 0.9970         112
#> 113 0.8610 0.9996 0.9737 0.9979         113
#> 114 0.8609 0.9997 0.9721 0.9984         114
#> 115 0.8608 0.9998 0.9663 0.9990         115
#> 116 0.8607 0.9999 0.9546 0.9996         116
#> 117 0.8607 0.9999 0.9528 0.9997         117
#> 118 0.8607 0.9999 0.9136 0.9999         118
#> 119 0.8607 1.0000 0.7373 1.0000         119
#> 120 0.8607 1.0000 0.0000 1.0000         120
#> 121 0.8647 0.9990 0.9866 0.9883         121
#> 122 0.8622 0.9992 0.9827 0.9926         122
#> 123 0.8616 0.9993 0.9797 0.9949         123
#> 124 0.8612 0.9995 0.9723 0.9971         124
#> 125 0.8610 0.9997 0.9741 0.9979         125
#> 126 0.8609 0.9997 0.9713 0.9985         126
#> 127 0.8608 0.9998 0.9672 0.9990         127
#> 128 0.8607 0.9999 0.9551 0.9996         128
#> 129 0.8607 0.9999 0.9527 0.9997         129
#> 130 0.8607 0.9999 0.9001 0.9999         130
#> 131 0.8607 1.0000 0.7149 1.0000         131
#> 132 0.8606 1.0000 0.0000 1.0000         132
#> 133 0.8646 0.9990 0.9864 0.9889         133
#> 134 0.8622 0.9992 0.9827 0.9929         134
#> 135 0.8615 0.9994 0.9798 0.9951         135
#> 136 0.8612 0.9995 0.9724 0.9972         136
#> 137 0.8610 0.9997 0.9741 0.9980         137
#> 138 0.8609 0.9997 0.9722 0.9985         138
#> 139 0.8608 0.9998 0.9619 0.9992         139
#> 140 0.8607 0.9999 0.9558 0.9996         140
#> 141 0.8607 0.9999 0.9520 0.9997         141
#> 142 0.8607 0.9999 0.8860 0.9999         142
#> 143 0.8606 1.0000 0.6675 1.0000         143
#> 144 0.8606 1.0000 0.0000 1.0000         144
#> 
#> Overall accuracy of the algorithm:
#>      PSe    PSp   PPPV   PNPV
#> 1 0.8629 0.9996 0.9811 0.9967
#> 
#> PSe denotes the pooling sensitivity.
#> PSp denotes the pooling specificity.
#> PPPV denotes the pooling positive predictive value.
#> PNPV denotes the pooling negative predictive value.

# Calculate the operating characteristics for
#   informative two-stage hierarchical testing
#   with a multiplex assay for two diseases.
config.mat <- matrix(data = c(rep(1, 5), rep(2, 4),
                              1, 1:10),
                     nrow = 2, ncol = 10, byrow = TRUE)
Se <- matrix(data = c(rep(0.95, 2), rep(0.99, 2)),
             nrow = 2, ncol = 2, byrow = FALSE)
Sp <- matrix(data = c(rep(0.96, 2), rep(0.98, 2)),
             nrow = 2, ncol = 2, byrow = FALSE)
calc3 <- opChar2(algorithm = "ID2",
                 alpha = c(18.25, 0.75, 0.75, 0.25),
                 Se = Se, Sp = Sp,
                 hier.config = config.mat)
#> 
#>  Number of minutes running:  0 
#>  
summary(calc3)
#> 
#> Algorithm: Informative two-stage hierarchical testing 
#> 
#> Testing configuration:
#> Block size: 10
#> Group sizes: 6,4
#> 
#> Expected number of tests: 5.29
#> Expected number of tests per individual: 0.5287
#> 
#> Disease 1 accuracy for individuals:
#>       PSe    PSp   PPPV   PNPV Individuals
#> 1  0.9639 0.9891 0.0133 1.0000           1
#> 2  0.9602 0.9869 0.0735 1.0000           2
#> 3  0.9630 0.9111 0.2269 0.9989           3
#> 4  0.9577 0.9274 0.2863 0.9986           4
#> 5  0.9569 0.9594 0.2699 0.9993           5
#> 6  0.9576 0.9793 0.6823 0.9980           6
#> 7  0.9724 0.9626 0.6694 0.9978           7
#> 8  0.9559 0.9887 0.2624 0.9998           8
#> 9  0.9646 0.9494 0.7321 0.9947           9
#> 10 0.9538 0.9561 0.3384 0.9989          10
#> 
#> Disease 2 accuracy for individuals:
#>       PSe    PSp   PPPV   PNPV Individuals
#> 1  0.9561 0.9830 0.2910 0.9997           1
#> 2  0.9561 0.9754 0.3793 0.9993           2
#> 3  0.9736 0.9536 0.1530 0.9998           3
#> 4  0.9555 0.9863 0.2232 0.9998           4
#> 5  0.9555 0.9683 0.4528 0.9987           5
#> 6  0.9545 0.9854 0.2780 0.9997           6
#> 7  0.9797 0.9626 0.5722 0.9989           7
#> 8  0.9559 0.8943 0.5529 0.9933           8
#> 9  0.9813 0.9682 0.5602 0.9992           9
#> 10 0.9552 0.8912 0.5423 0.9933          10
#> 
#> Overall accuracy of the algorithm:
#>      PSe    PSp   PPPV   PNPV
#> 1 0.9634 0.9613 0.4683 0.9987
#> 2 0.9615 0.9582 0.4866 0.9983
#> 
#> PSe denotes the pooling sensitivity.
#> PSp denotes the pooling specificity.
#> PPPV denotes the pooling positive predictive value.
#> PNPV denotes the pooling negative predictive value.

# Calculate the operating characteristics for
#   non-informative array testing with master pooling
#   with a multiplex assay for two diseases.
calc4 <- opChar2(algorithm = "A2M",
                 p.vec = c(0.92, 0.05, 0.02, 0.01),
                 Se = rep(0.95, 2), Sp = rep(0.99, 2),
                 rowcol.sz = 8)
#> 
#>  Number of minutes running:  0 
#>  
summary(calc4)
#> 
#> Algorithm: Non-informative array testing with master pooling 
#> 
#> Testing configuration:
#> Row/column size: 8
#> Array size: 64
#> 
#> Expected number of tests: 30.29
#> Expected number of tests per individual: 0.4733
#> 
#> Disease 1 accuracy for individuals:
#>      PSe    PSp   PPPV   PNPV Individuals
#> 1 0.8704 0.9983 0.9709 0.9918         All
#> 
#> Disease 2 accuracy for individuals:
#>      PSe    PSp   PPPV   PNPV Individuals
#> 1 0.9004 0.9981 0.9366 0.9969         All
#> 
#> Overall accuracy of the algorithm:
#>      PSe    PSp   PPPV   PNPV
#> 1 0.8704 0.9983 0.9709 0.9918
#> 2 0.9004 0.9981 0.9366 0.9969
#> 
#> PSe denotes the pooling sensitivity.
#> PSp denotes the pooling specificity.
#> PPPV denotes the pooling positive predictive value.
#> PNPV denotes the pooling negative predictive value.